refactor: move power files into power/ module directory
Groups power_monitor.rs, power_profile.rs, and power_manager.rs under
a single power/ module with clean re-exports via mod.rs. Internal
imports use `super::` and external consumers use `crate::power::`.
