+++ crates/screenpipe-server/src/lib.rs
@@ -25,9 +25,7 @@ pub mod hot_frame_cache;
 pub mod meeting_persister;
 pub mod meeting_watcher;
 pub mod paired_capture;
-pub mod power_manager;
-pub mod power_monitor;
-pub mod power_profile;
+pub mod power;
 mod resource_monitor;
 mod server;
 pub mod sleep_monitor;
@@ -57,7 +55,7 @@ pub use server::HealthCheckResponse;
 pub use server::PaginatedResponse;
 pub use server::SCServer;
 pub use server::{api_list_monitors_handler as api_list_monitors, MonitorInfo};
-pub use power_manager::{start_power_manager, PowerManagerHandle};
+pub use power::{start_power_manager, PowerManagerHandle};
 pub use sleep_monitor::start_sleep_monitor;
 pub use ui_recorder::{start_ui_recording, UiRecorderConfig, UiRecorderHandle};
 pub use video::{video_quality_to_crf, video_quality_to_jpeg_q, video_quality_to_preset};
