+++ crates/screenpipe-server/src/power/monitor.rs
